Thermocouple temperature is calculated using junection compensation rather than just adding junction temperature and thermocouple temperature. Added an MCP3424 compensation factory that gives really good results. Estimating junction voltage, TODO implement this properly.

+++ t400/functions.h
@@ -68,7 +68,7 @@ namespace Display {
 // Converts the junction temperature into a voltage for offset
 // @param temperature reading from junction temperature sensor
 // return voltage in uVolts
-double GetJunctionVoltage(uint16_t jTemp);
+int32_t GetJunctionVoltage(double jTemp);
 
 // Converts the thermocouple µV reading into some usable °C
 // @param microVolt reading from the ADC
